override headers with a func
