+++ src/MudBlazor.Markdown/MudMarkdown.razor.cs
@@ -42,53 +42,18 @@ namespace MudBlazor
 		public ICommand? LinkCommand { get; set; }
 
 		/// <summary>
-		/// Typography variant to use for Heading Level 1.<br/>
-		/// Default: <see cref="Typo.h1"/>
-		/// </summary>
-		[Parameter]
-		public Typo H1Typo { get; set; } = Typo.h1;
-
-		/// <summary>
-		/// Typography variant to use for Heading Level 2.<br/>
-		/// Default: <see cref="Typo.h2"/>
-		/// </summary>
-		[Parameter]
-		public Typo H2Typo { get; set; } = Typo.h2;
-
-		/// <summary>
-		/// Typography variant to use for Heading Level 3.<br/>
-		/// Default: <see cref="Typo.h3"/>
-		/// </summary>
-		[Parameter]
-		public Typo H3Typo { get; set; } = Typo.h3;
-
-		/// <summary>
-		/// Typography variant to use for Heading Level 4.<br/>
-		/// Default: <see cref="Typo.h4"/>
-		/// </summary>
-		[Parameter]
-		public Typo H4Typo { get; set; } = Typo.h4;
-
-		/// <summary>
-		/// Typography variant to use for Heading Level 5.<br/>
-		/// Default: <see cref="Typo.h5"/>
-		/// </summary>
-		[Parameter]
-		public Typo H5Typo { get; set; } = Typo.h5;
-
-		/// <summary>
-		/// Typography variant to use for Heading Level 6.<br/>
-		/// Default: <see cref="Typo.h6"/>
-		/// </summary>
-		[Parameter]
-		public Typo H6Typo { get; set; } = Typo.h6;
-
-		/// <summary>
 		/// Override the original URL address of the <see cref="LinkInline"/>.<br/>
-		/// If the parameter is not provider <see cref="LinkInline"/>.<see cref="LinkInline.Url"/> is used
+		/// If a function is not provided <see cref="LinkInline.Url"/> is used
 		/// </summary>
 		[Parameter]
 		public Func<LinkInline, string?>? OverrideLinkUrl { get; set; }
+
+		/// <summary>
+		/// Typography variant to use for Heading Level 1-6.<br/>
+		/// If a function is not provided a default typo for each level is set (e.g. for &lt;h1&gt; it will be <see cref="Typo.h1"/>, etc.)
+		/// </summary>
+		[Parameter]
+		public Func<Typo, Typo>? OverrideHeaderTypo { get; set; }
 
 		[Inject]
 		private NavigationManager? NavigationManager { get; init; }
@@ -159,16 +124,8 @@ namespace MudBlazor
 						}
 					case HeadingBlock heading:
 						{
-							Typo? typo = heading.Level switch
-							{
-								1 => H1Typo,
-								2 => H2Typo,
-								3 => H3Typo,
-								4 => H4Typo,
-								5 => H5Typo,
-								6 => H6Typo,
-								_ => null
-							};
+							Typo? typo = (Typo)heading.Level;
+							typo = OverrideHeaderTypo?.Invoke((Typo)heading.Level) ?? typo;
 
 							if (typo.HasValue)
 							{
