+++ tests/MudBlazor.Markdown.Tests/MarkdownComponentTests/MarkdownComponentTestsBase.cs
@@ -13,8 +13,7 @@ namespace MudBlazor.Markdown.Tests.MarkdownComponentTests
 		protected IRenderedComponent<MudMarkdown> CreateFixture(
 			string value,
 			Optional<ICommand> command = default, Optional<int?> tableCellMinWidth = default,
-			Optional<Func<LinkInline, string>> overrideLinkUrl = default,
-			Optional<Typo> h1Typo = default, Optional<Typo> h2Typo = default, Optional<Typo> h3Typo = default, Optional<Typo> h4Typo = default, Optional<Typo> h5Typo = default, Optional<Typo> h6Typo = default)
+			Optional<Func<LinkInline, string>> overrideLinkUrl = default, Optional<Func<Typo, Typo>> overrideHeaderTypo = default)
 		{
 			MockNavigationManager.Initialize(Uri);
 
@@ -23,12 +22,7 @@ namespace MudBlazor.Markdown.Tests.MarkdownComponentTests
 					.TryAdd(static x => x.LinkCommand, command)
 					.TryAdd(static x => x.TableCellMinWidth, tableCellMinWidth)
 					.TryAdd(static x => x.OverrideLinkUrl, overrideLinkUrl)
-					.TryAdd(static x => x.H1Typo, h1Typo)
-					.TryAdd(static x => x.H2Typo, h2Typo)
-					.TryAdd(static x => x.H3Typo, h3Typo)
-					.TryAdd(static x => x.H4Typo, h4Typo)
-					.TryAdd(static x => x.H5Typo, h5Typo)
-					.TryAdd(static x => x.H6Typo, h6Typo));
+					.TryAdd(static x => x.OverrideHeaderTypo, overrideHeaderTypo));
 		}
 	}
 }
